Added facenet impl\n Fix bug: size x4

+++ preprocess/image_preprocess.py
@@ -2,6 +2,7 @@ import numpy as np
 import sys
 import dlib
 import cv2
+import os
 from os.path import join, basename, splitext
 
 import constants 
@@ -27,8 +28,11 @@ def __align(image, landmarks, landmarkIndices, output_dimension):
 
 def get_faces(image_path, output_dimension=128, landmarkIndices=constants.OUTER_EYES_AND_NOSE):
 	image = cv2.imread(image_path)
-
-	detected_faces = __face_detector(image, 1)
+	try:
+		detected_faces = __face_detector(image, 1)
+	except:
+		print('Failed at image:', image_path)
+		return []
 	#print("detected faces", len(detected_faces))
 	
 	images = []
@@ -42,18 +46,26 @@ def get_faces(image_path, output_dimension=128, landmarkIndices=constants.OUTER_
 
 
 def save_faces(image_path, output_folder, output_dimension=128, landmarkIndices=constants.OUTER_EYES_AND_NOSE, extension='.png'):
+	base, ext = splitext(basename(image_path))
+	if extension is None: 
+		extension = ext
+
+	def get_img_path(id):
+		new_image_name = "{}_{}{}".format(base, id, extension)
+		return join(output_folder, new_image_name)
+
+
+	first_image_path = get_img_path(0)
+	if os.path.exists(first_image_path):
+		print('Skipping existing ', image_path)
+		return []
+
+
 	images = get_faces(image_path, output_dimension, landmarkIndices)
-	
 	image_paths = []
 	for i, image in enumerate(images):
-		base, ext = splitext(basename(image_path))
-		if extension is None: 
-			extension = ext
-		
-		new_image_name = "{}_{}{}".format(base, i, extension)
-		new_image_path = join(output_folder, new_image_name)
+		new_image_path = get_img_path(i)
 		cv2.imwrite(new_image_path, image) 
-		
 		image_paths.append(new_image_path)
 
 	return image_paths
